requirements-dev.txt
ray[tune]==1.12.1
fiftyone==0.16.0
lightning-flash[image, text]==0.7.5
click==8
streamlit==1.9.0
jinja2==3.0.3
gradio==3.0.5
protobuf==3.20.1
